Implement multilingual hierarchical search links
Replace `%entity-name%` and `%url-name%` by `url-target` for url only.

New search links aim to replace default autolinks as follows:
- Reorganize links in hierarchical sections.
- Autohide empty sections (fix #206 issue).
- Support section show/hide configuration.
- Support section title translation.
- Support link title/target localization.
- Display each search links category apart.

The proposed hierarchical organization is:
- (general-purpose/music-agnostic) web search links
- links to search the MB URL relationships whitelist only
  - lyrics DBs
  - score DBs
  - regional DBs
  - genre DBs
  - other DBs
- links to other specific searches
  - (same subcategories as above)
  - (maybe more)

Example links and sections are provided (regardless of charsets).

This version can be used to compare search links to default autolinks.
Note that search links configuration and default autolinks one are
independent, so that it is always possible for users to step back.